pylint and permisions

+++ tournament/views.py
@@ -13,6 +13,7 @@ from league.models import User, Sgf
 from league.forms import SgfAdminForm, ActionForm
 from fullcalendar.forms import UTCPublicEventForm
 from community.forms import CommunytyUserForm
+from pytz import utc
 from .models import Tournament, Bracket, Match, TournamentPlayer, TournamentGroup, Round, TournamentEvent
 from .forms import TournamentForm, TournamentGroupForm, RoundForm, TournamentAboutForm, TournamentPlayerProfileForm
 from .utils import save_round
@@ -128,7 +129,7 @@ def calendar(request, tournament_id):
     template = loader.get_template('tournament/calendar.html')
     return HttpResponse(template.render(context, request))
 
-def calendar_feed(request,tournament_id):
+def calendar_feed(request, tournament_id):
     """get all events for one user and serve a json."""
 
     tournament = get_object_or_404(Tournament, pk=tournament_id)
@@ -281,10 +282,10 @@ def tournament_list(request):
 
 
 @login_required()
-@user_passes_test(User.is_league_admin, login_url="/", redirect_field_name=None)
 def manage_settings(request, tournament_id):
     tournament = get_object_or_404(Tournament, pk=tournament_id)
-
+    if not tournament.is_admin(request.user):
+        raise Http404('What are you doing here?')
     if request.method == 'POST':
         form = TournamentForm(request.POST, instance=tournament)
         if form.is_valid:
@@ -302,9 +303,10 @@ def manage_settings(request, tournament_id):
     return HttpResponse(template.render(context, request))
 
 @login_required()
-@user_passes_test(User.is_league_admin, login_url="/", redirect_field_name=None)
 def manage_groups(request, tournament_id):
     tournament = get_object_or_404(Tournament, pk=tournament_id)
+    if not tournament.is_admin(request.user):
+        raise Http404('What are you doing here?')
     players = TournamentPlayer.objects.filter(event=tournament).order_by('order')
     groups = TournamentGroup.objects.filter(league_event=tournament).order_by('order')
     context = {
@@ -316,9 +318,10 @@ def manage_groups(request, tournament_id):
     return HttpResponse(template.render(context, request))
 
 @login_required()
-@user_passes_test(User.is_league_admin, login_url="/", redirect_field_name=None)
 def create_bracket(request, tournament_id):
     tournament = get_object_or_404(Tournament, pk=tournament_id)
+    if not tournament.is_admin(request.user):
+        raise Http404('What are you doing here?')
     if request.method == 'POST':
         bracket = Bracket(tournament=tournament)
         bracket.order = tournament.last_bracket_order() + 1
@@ -329,9 +332,10 @@ def create_bracket(request, tournament_id):
     ))
 
 @login_required()
-@user_passes_test(User.is_league_admin, login_url="/", redirect_field_name=None)
 def create_match(request, round_id):
     round = get_object_or_404(Round, pk=round_id)
+    if not round.bracket.tournament.is_admin(request.user):
+        raise Http404('What are you doing here?')
     if request.method == 'POST':
         round.create_match()
 
@@ -341,10 +345,11 @@ def create_match(request, round_id):
     ))
 
 @login_required()
-@user_passes_test(User.is_league_admin, login_url="/", redirect_field_name=None)
 def delete_match(request, round_id):
     """Delete the last match of a tournament"""
     round = get_object_or_404(Round, pk=round_id)
+    if not round.bracket.tournament.is_admin(request.user):
+        raise Http404('What are you doing here?')
     match = round.match_set.all().order_by('order').last()
     if match is not None:
         if match.player_1 or match.player_2:
@@ -358,9 +363,10 @@ def delete_match(request, round_id):
     ))
 
 @login_required()
-@user_passes_test(User.is_league_admin, login_url="/", redirect_field_name=None)
 def rename_round(request, round_id):
     round = get_object_or_404(Round, pk=round_id)
+    if not round.bracket.tournament.is_admin(request.user):
+        raise Http404('What are you doing here?')
     if request.method == 'POST':
         form = RoundForm(request.POST)
         if form.is_valid():
@@ -373,9 +379,10 @@ def rename_round(request, round_id):
 
 
 @login_required()
-@user_passes_test(User.is_league_admin, login_url="/", redirect_field_name=None)
 def create_round(request, bracket_id):
     bracket = get_object_or_404(Bracket, pk=bracket_id)
+    if not bracket.tournament.is_admin(request.user):
+        raise Http404('What are you doing here?')
     if request.method == 'POST':
         form = RoundForm(request.POST)
         if form.is_valid():
@@ -394,9 +401,10 @@ def create_round(request, bracket_id):
 
 
 @login_required()
-@user_passes_test(User.is_league_admin, login_url="/", redirect_field_name=None)
 def rename_bracket(request, bracket_id):
     bracket = get_object_or_404(Bracket, pk=bracket_id)
+    if not bracket.tournament.is_admin(request.user):
+        raise Http404('What are you doing here?')
     if request.method == 'POST':
         form = RoundForm(request.POST)
         if form.is_valid():
@@ -409,9 +417,10 @@ def rename_bracket(request, bracket_id):
     ))
 
 @login_required()
-@user_passes_test(User.is_league_admin, login_url="/", redirect_field_name=None)
 def delete_bracket(request, bracket_id):
     bracket = get_object_or_404(Bracket, pk=bracket_id)
+    if not bracket.tournament.is_admin(request.user):
+        raise Http404('What are you doing here?')
     if request.method == 'POST':
         bracket.delete()
     return HttpResponseRedirect(reverse(
@@ -421,9 +430,10 @@ def delete_bracket(request, bracket_id):
 
 
 @login_required()
-@user_passes_test(User.is_league_admin, login_url="/", redirect_field_name=None)
 def delete_round(request, round_id):
     round = get_object_or_404(Round, pk=round_id)
+    if not round.bracket.tournament.is_admin(request.user):
+        raise Http404('What are you doing here?')
     if request.method == 'POST':
         round.delete()
     return HttpResponseRedirect(reverse(
@@ -433,9 +443,10 @@ def delete_round(request, round_id):
 
 
 @login_required()
-@user_passes_test(User.is_league_admin, login_url="/", redirect_field_name=None)
 def save_brackets(request, tournament_id):
     tournament = get_object_or_404(Tournament, pk=tournament_id)
+    if not tournament.is_admin(request.user):
+        raise Http404('What are you doing here?')
     if request.method == 'POST':
         brackets = json.loads(request.POST.get('brackets'))
         for bracket_id, rounds in brackets.items():
@@ -448,9 +459,10 @@ def save_brackets(request, tournament_id):
 
 
 @login_required()
-@user_passes_test(User.is_league_admin, login_url="/", redirect_field_name=None)
 def manage_brackets(request, tournament_id):
     tournament = get_object_or_404(Tournament, pk=tournament_id)
+    if not tournament.is_admin(request.user):
+        raise Http404('What are you doing here?')
     players = TournamentPlayer.objects.filter(event=tournament).order_by('order')
     brackets = tournament.bracket_set.all()
     if not brackets:
@@ -478,10 +490,11 @@ def manage_brackets(request, tournament_id):
     return HttpResponse(template.render(context, request))
 
 @login_required()
-@user_passes_test(User.is_league_admin, login_url="/", redirect_field_name=None)
 def manage_games(request, tournament_id):
     """Manage tournament games."""
     tournament = get_object_or_404(Tournament, pk=tournament_id)
+    if not tournament.is_admin(request.user):
+        raise Http404('What are you doing here?')
     games = Sgf.objects.filter(events=tournament)
     if request.method == 'POST':
         pass
@@ -494,10 +507,11 @@ def manage_games(request, tournament_id):
 
 
 @login_required()
-@user_passes_test(User.is_league_admin, login_url="/", redirect_field_name=None)
 def set_stage(request, tournament_id):
     """Set the stage of a tournament."""
     tournament = get_object_or_404(Tournament, pk=tournament_id)
+    if not tournament.is_admin(request.user):
+        raise Http404('What are you doing here?')
     if request.method == 'POST':
         form = ActionForm(request.POST)
         if form.is_valid():
@@ -510,10 +524,11 @@ def set_stage(request, tournament_id):
 
 
 @login_required()
-@user_passes_test(User.is_league_admin, login_url="/", redirect_field_name=None)
 def create_sgf(request, tournament_id):
     """Actually create a sgf db entry. Should be called after upload_sgf."""
     tournament = get_object_or_404(Tournament, pk=tournament_id)
+    if not tournament.is_admin(request.user):
+        raise Http404('What are you doing here?')
     if request.method == 'POST':
         form = SgfAdminForm(request.POST)
         if form.is_valid():
@@ -547,12 +562,13 @@ def create_sgf(request, tournament_id):
     ))
 
 @login_required()
-@user_passes_test(User.is_league_admin, login_url="/", redirect_field_name=None)
 def upload_sgf(request, tournament_id):
     """THis view allow user to preview sgf with wgo along with valid status of the sgf.
         Can call save_sgf from it.
     """
     tournament = get_object_or_404(Tournament, pk=tournament_id)
+    if not tournament.is_admin(request.user):
+        raise Http404('What are you doing here?')
     if request.method == 'POST':
         form = SgfAdminForm(request.POST)
         if form.is_valid():
@@ -595,10 +611,11 @@ def upload_sgf(request, tournament_id):
             raise Http404("What are you doing here ?")
 
 @login_required()
-@user_passes_test(User.is_league_admin, login_url="/", redirect_field_name=None)
 def invite_user(request, tournament_id):
     """Invite a user in a tournament."""
     tournament = get_object_or_404(Tournament, pk=tournament_id)
+    if not tournament.is_admin(request.user):
+        raise Http404('What are you doing here?')
     if request.method == 'POST':
         form = CommunytyUserForm(request.POST)
         if form.is_valid():
@@ -632,11 +649,12 @@ def invite_user(request, tournament_id):
         ))
 
 @login_required()
-@user_passes_test(User.is_league_admin, login_url="/", redirect_field_name=None)
 def remove_players(request, tournament_id):
     """Remove a player from a tournament ajax powa"""
+    tournament = get_object_or_404(Tournament, pk=tournament_id)
+    if not tournament.is_admin(request.user):
+        raise Http404('What are you doing here?')
     if request.method == "POST":
-        tournament = get_object_or_404(Tournament, pk=tournament_id)
         players_list = json.loads(request.POST.get('players_list'))
         players = TournamentPlayer.objects.filter(pk__in=players_list)
         players.delete()
@@ -645,10 +663,11 @@ def remove_players(request, tournament_id):
         raise Http404('what are you doing here ?')
 
 @login_required()
-@user_passes_test(User.is_league_admin, login_url="/", redirect_field_name=None)
 def save_players_order(request, tournament_id):
     """Save the player order and remove players from ajax call"""
     tournament = get_object_or_404(Tournament, pk=tournament_id)
+    if not tournament.is_admin(request.user):
+        raise Http404('What are you doing here?')
     if request.method == 'POST':
         players_list = json.loads(request.POST.get('players_list'))
         # Now we update the players order
@@ -663,9 +682,10 @@ def save_players_order(request, tournament_id):
 
 
 @login_required()
-@user_passes_test(User.is_league_admin, login_url="/", redirect_field_name=None)
 def create_group(request, tournament_id):
     tournament = get_object_or_404(Tournament, pk=tournament_id)
+    if not tournament.is_admin(request.user):
+        raise Http404('What are you doing here?')
     if request.method == 'POST':
         form = TournamentGroupForm(request.POST)
         if form.is_valid():
@@ -679,10 +699,11 @@ def create_group(request, tournament_id):
 
 
 @login_required()
-@user_passes_test(User.is_league_admin, login_url="/", redirect_field_name=None)
 def save_groups(request, tournament_id):
     """Save tournament groups players from ajax call."""
     tournament = get_object_or_404(Tournament, pk=tournament_id)
+    if not tournament.is_admin(request.user):
+        raise Http404('What are you doing here?')
     if request.method == 'POST':
         # first we null all players division
         players = TournamentPlayer.objects.filter(event=tournament)
